220601_StaticFusion_TDoA_Manual_and_TB
MATLAB scripts that were curated in the very initial stages of working on the TDoA algorithms for a passive multi-lateration system. These scripts were mostly used to build up an understanding and test the theory of the TDoA localization principle. This was carried out using the MATLAB Sensor Fusion Toolbox which utilizes the Static Fusion algorithm to cooperate with the out-of-sequence measurements (OOSM) of the Time of Arrival (ToA) data and aligns them in time for coherency in TDoA measurements for further processing. 

This was attempted manually as well as using an example built in MATLAB