Summary of GDoP implementation and math model
This document is a brief tutorial into the mathematics, steps and algorithmic implementation considerations that were taken into account while implementing the GDoP algorithm. It is in no way a very comprehensive document, but it gives an educated reader some idea about whats going on. There may be shortcomings, assumptions and mistakes, but the math is correct and referenced from reliable, published sources. I am open to any suggestions and recommendations.