Statistical performance evaluation of TDoA
The main script: 'tdoaGDoP_bench_250110' is usedto create a user defined scenario with a custom RMS timing error, sensor and emitter positions, no of monte Carli iterations and other hyperparatmers to adjust the simulation per need, i.e., confidence interval for the error ellipse etc.

The TDoA based localization is carried out in the presence of zero mean Gaussian distributed noise. To gauge whether the estimates are consistent and robust, and to get an idea of the bias, the GDoP is calculated for the emitter position and the given network placement. The Circular Error Probability and Error ellipse are also computed using the measurement covariance or the CRLB. 

The mean GDoP is compared with the Bias at each iteration.
The CEP50 and error ellipse are computed for the estimator/solver using the estimate position error covariance. This gives a visual and quantified insight into how well the estimator/solver is performing.