Add files via upload
Verification of the synergy of the polynomial hyperbolic solution obtained from solving the non-linear equations of hyperbola and that of the iterative least squares method such that the final estimate of the emitter location is improved significantly compared to the analytical-only solution. Fundamentally, once, based on the TDoA measurements, the emitter location is estimated, it is passed on as an 'initial guess' along with information and model of the noise and sensor geometry. The iterative least squares method then minimizes the (vertical) residual between the estimated RDoA and the measured RDoAs until the tolerance achieved of the solution begins to diverge. 
The following functions are a part of the entire code:
1. fn_tdoaOLS: The main function that computes the 'Ordinary Least Squares' 
2. fn_tdoaGen: This function generates the TOA measurements at each sensor based on the ranges of the emitter from the sensor. Gaussian noise is added to the measurements
3. fn_RDoAErrCov: Computes the error covariance for the given timing error based on a centralized TDoA computation architecture
4. fn_modgdop2D: Computes the GDoP for the given scenario. Also outputs the FIM.