BruteForce_FDoA_CoarseFine_GridSearch
This is a two-step solver. The process is divided into two stages where the estimation of the position of the emitter is first calculated with an intial guess for the emitters 
 velocity. With the variables reduced down to 3 (x,y,z) instead of 6 (x,y,z,vx,vy,vz), a CRS-based Solution of the 3 baseline FDoA equations yield the solution which is the location of the emitter. Once the position has been estimated, the solution can be used again by susbstituting the position to determine the velocity of the emitter.
This version of the script incorporates Adaptive Grid Refinement and the iterative Quasi-Newton optimization method for FDoA-based emitter localization. The goal is to reduce computational overhead while maintaining accuracy. The grid search method is divided into two steps where the search is first carried out across the wider span with a lower resolution which isolates the region of interest without compromising on the timing and computational load. Based on the coarse estimate obtained, a finer grid search is initiated in a smaller region.